# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Prisioner API Documentation'
  description: "Documentação da API para o Sistema de Gerenciamento de Prisioneiros\n    Esta documentação tem como objetivo fornecer todas as informações necessárias para trabalhar com nossa API.\n\n    Conforme você rola, verá exemplos de código para trabalhar com a API em diferentes linguagens de programação \n    na área escura à direita (ou como parte do conteúdo em dispositivos móveis).\n\n    Você pode alternar a linguagem usada com as guias no canto superior direito (ou no menu de navegação no canto \n    superior esquerdo em dispositivos móveis)."
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
  /api/users:
    get:
      summary: ''
      operationId: getApiUsers
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    post:
      summary: ''
      operationId: postApiUsers
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
  '/api/users/{id}':
    delete:
      summary: ''
      operationId: deleteApiUsersId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the user.'
        example: sed
        required: true
        schema:
          type: string
  /api/employees:
    get:
      summary: ''
      operationId: getApiEmployees
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    post:
      summary: ''
      operationId: postApiEmployees
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
  '/api/employees/{id}':
    delete:
      summary: ''
      operationId: deleteApiEmployeesId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the employee.'
        example: perspiciatis
        required: true
        schema:
          type: string
  /api/prisioners:
    get:
      summary: ''
      operationId: getApiPrisioners
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Success
                properties:
                  message:
                    type: string
                    example: Success
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Forbidden
                properties:
                  message:
                    type: string
                    example: Forbidden
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Not Found'
                properties:
                  message:
                    type: string
                    example: 'Not Found'
        405:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Method Not Allowed'
                properties:
                  message:
                    type: string
                    example: 'Method Not Allowed'
      tags:
        - Endpoints
      security: []
    post:
      summary: ''
      operationId: postApiPrisioners
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
  '/api/prisioners/{id}':
    delete:
      summary: ''
      operationId: deleteApiPrisionersId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the prisioner.'
        example: inventore
        required: true
        schema:
          type: string
  /api/visitors:
    get:
      summary: ''
      operationId: getApiVisitors
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    post:
      summary: ''
      operationId: postApiVisitors
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
  '/api/visitors/{id}':
    delete:
      summary: ''
      operationId: deleteApiVisitorsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the visitor.'
        example: perferendis
        required: true
        schema:
          type: string
tags:
  -
    name: Endpoints
    description: ''